+++ provider/provider.go
@@ -85,6 +85,7 @@ func (p *pangolinProvider) Configure(ctx context.Context, req provider.Configure
 
 func (p *pangolinProvider) Resources(_ context.Context) []func() resource.Resource {
 	return []func() resource.Resource{
+		NewPangolinSiteResource,
 		NewSiteResource,
 		NewTargetResource,
 		NewRoleResource,